% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmprobe-package.R
\docType{package}
\name{lmmprobe-package}
\alias{lmmprobe-package}
\title{lmmprobe: Sparse High-Dimensional Linear Mixed Modeling with a Partitioned Empirical Bayes ECM Algorithm}
\description{
Implements a partitioned Empirical Bayes ECM algorithm for sparse high-dimensional linear mixed modeling. The package provides efficient estimation and inference for mixed models with high-dimensional fixed effects.
}
\author{
\strong{Maintainer}: Anja Zgodic \email{azgodic@email.sc.edu}

Authors:
\itemize{
  \item Ray Bai (\href{https://orcid.org/0000-0001-7150-7229}{ORCID})
  \item Jiajia Zhang (\href{https://orcid.org/0000-0001-8281-8532}{ORCID})
  \item Alex McLain (\href{https://orcid.org/0000-0002-5475-0670}{ORCID})
  \item Peter Olejua (\href{https://orcid.org/0000-0002-2478-0908}{ORCID})
}

}
\keyword{internal}
