% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SLE}
\alias{SLE}
\alias{real_data}
\title{Systemic Lupus Erythematosus (SLE) Data}
\format{
A data frame with observations on the following variables:
\describe{
  \item{y}{Response vector.}
  \item{id}{Subject ID.}
  \item{time}{Time of observation.}
  \item{...}{Other predictors.}
}
}
\source{
Simulated or real data provided with the package.
}
\usage{
data(SLE)
}
\description{
A dataset containing data from a study on Systemic Lupus Erythematosus (SLE).
Loading this dataset creates an object named `real_data`.
}
