\name{lmmprobe-package}
\alias{lmmprobe-package}
\alias{lmmprobe}
\docType{package}
\title{\packageTitle{lmmprobe}}
\description{\packageDescription{lmmprobe}}
\section{Package Content}{\packageIndices{lmmprobe}}
\author{\packageAuthor{lmmprobe}}
\section{Maintainer}{\packageMaintainer{lmmprobe}}
\keyword{package}
