% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmprobe.R
\name{lmmprobe}
\alias{lmmprobe}
\title{Sparse high-dimensional linear mixed modeling with PaRtitiOned empirical Bayes ECM (LMM-PROBE) algorithm.}
\arguments{
\item{Y}{A training-data matrix containing the outcome \code{Y}.}

\item{Z}{A training-data matrix containing the sparse fixed-effect predictors on which to apply the lmmprobe algorithm. The first columns should be the "id" column.}

\item{V}{A training-data matrix containing non-sparse predictors for the random effects. This matrix is currently only programmed for two scenarios. Scenario 1: only a random intercept, where V is a matrix with one column containing ID's and each unit has the same number of observations. Scenario 2: a random intercept and a random slope, where V is a matrix with two columns. The first column is ID and the second column is a continuous variable (e.g. time) for which a random slope is to be estimated. Each unit has the same number of observations.}

\item{ID_data}{A factor vector of IDs for subjects in the training set.}

\item{Y_test}{A testing-data matrix containing the outcome \code{Y}. Default is NULL.}

\item{Z_test}{A training-data matrix containing the sparse fixed-effect predictors on which to apply the lmmprobe algorithm. The first columns should be the "id" column. Default is NULL.}

\item{V_test}{A training-data matrix containing non-sparse predictors for the random effects. This matrix is currently only programmed for two scenarios. Scenario 1: only a random intercept, where V is a matrix with one column containing ID's and each unit has the same number of observations. Scenario 2: a random intercept and a random slope, where V is a matrix with two columns. The first column is ID and the second column is a continuous variable (e.g. time) for which a random slope is to be estimated. Each unit has the same number of observations. Default is NULL.}

\item{ID_test}{A factor vector of IDs for subjects in the testing set.}

\item{alpha}{Type I error; significance level.}

\item{ep}{Value against which to compare convergence criterion, we recommend 0.05.}

\item{B}{The number of groups to categorize estimated coefficients in to calculate correlation \eqn{\rho}. We recommend five.}

\item{adj}{A factor multiplying Silverman’s 'rule of thumb' in determining the bandwidth for density estimation, same as the 'adjust' argument of R's density function. Default is three.}

\item{maxit}{Maximum number of iterations the algorithm will run for. Default is 10000.}

\item{cpus}{The number of CPUS user would like to use for parallel computations. Default is four.}

\item{LR}{A learning rate parameter \code{r}. Using zero corresponds to the implementation described in Zgodic et al.}

\item{C}{A learning rate parameter \code{c}. Using one corresponds to the implementation described in Zgodic et al.}

\item{sigma_init}{An initial value for the residual variance parameter. Default is NULL which corresponds to the sample variance of Y.}
}
\value{
A list of the output of the lmmprobe function, including 

\code{beta_hat, beta_hat_var} MAP estimates of the posterior expectation (beta_hat) and variance (beta_hat_var) of the prior mean (\eqn{\beta}) of the regression coefficients assuming \eqn{\gamma=1}, 

\code{gamma} the posterior expectation of the latent \eqn{\gamma} variables, 

\code{preds} predictions of \eqn{Y},

\code{PI_lower, PI_upper} lower and upper prediction intervals for the predictions,

\code{sigma2_est} MAP estimate of the residual variance, 

\code{random_var} MAP estimate of the random effect(s) variance, 

\code{random_intercept} estimated random intercept terms,  

\code{random_slope} estimated random slope terms, if applicable.
}
\description{
Sparse high-dimensional linear mixed modeling with PaRtitiOned empirical Bayes ECM (LMM-PROBE) algorithm. Currently, the package offers functionality for two scenarios. Scenario 1: only a random intercept, each unit has the same number of observations; Scenario 2: a random intercept and a random slope, each unit has the same number of observations. We are actively expanding the package for more flexibility and scenarios.
}
\examples{
library(lmmprobe)
data(SLE)
Y <- matrix(real_data[,"y"], ncol=1)
Z <- real_data[,4:ncol(real_data)]
V <- matrix(real_data[,"id"], ncol=1)
ID_data <- as.numeric(as.character(real_data$id))
full_res <- lmmprobe(Y = Y, Z = Z, V = V, ID_data = ID_data)
}
\references{
Zgodic A, Bay R, Zhang J, Olejua P, and McLain AC (2023). Sparse high-dimensional linear mixed modeling with a partitioned empirical Bayes ECM algorithm. arXiv preprint arXiv:2310.12285.
}
